+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The Homebrew package has migrated from a formula to a cask. Install it with `brew install --cask reviewdog/tap/reviewdog`.
 
 ### :rocket: Enhancements
+
+### :bug: Fixes
+
+### :rotating_light: Breaking changes
+
+## [v0.21.1] - 2026-06-12
+
+### :rocket: Enhancements
 - [#2257](https://github.com/reviewdog/reviewdog/pull/2257) Use -trimpath flag for reproducible build
 - [#2513](https://github.com/reviewdog/reviewdog/pull/2513) Drop support of windows/arm
 - [#2543](https://github.com/reviewdog/reviewdog/pull/2543) Update `gitlab-mr-discussion` reporter to embed a fingerprint meta-comment in each posted note and automatically resolve previously-posted discussions whose diagnostic is no longer reported (fixes [#1150](https://github.com/reviewdog/reviewdog/issues/1150)).
@@ -322,7 +330,8 @@ $ cd subdir/ && reviewdog -filter-mode=file -fail-on-error -reporter=github-pr-r
 
 See https://github.com/reviewdog/reviewdog/releases for older release note.
 
-[Unreleased]: https://github.com/reviewdog/reviewdog/compare/v0.21.0...HEAD
+[Unreleased]: https://github.com/reviewdog/reviewdog/compare/v0.21.1...HEAD
+[v0.21.1]: https://github.com/reviewdog/reviewdog/compare/v0.21.0...v0.21.1
 [v0.21.0]: https://github.com/reviewdog/reviewdog/compare/v0.20.3...v0.21.0
 [v0.10.0]: https://github.com/reviewdog/reviewdog/compare/v0.9.17...v0.10.0
 [v0.10.1]: https://github.com/reviewdog/reviewdog/compare/v0.10.0...v0.10.1